refact/EasterEGG Modify
